Standardization - All compiles for me no issues here.

+++ src/main.cpp
@@ -1,8 +1,10 @@
-
-//asteroids.cpp
-//CS 202 Final Proj
-//Auth: William Horn
-//Rough draft for game framework
+// main.cpp
+// CS 202 Project: Scrolling Space Shooter
+// Team Members: William Horn, Corey Gray, Michael Bilan, Cameron Titus, Kyle Tam, Andrew Cummins
+// Created: 201704??
+// Updated: 20170426
+//
+// Game framework
 
 #include "constants.h"
 #include "Screens.h"
@@ -16,29 +18,35 @@ using sf::Color;
 #include <vector>
 using std::vector;
 
-int main(){
-    //vector of screens, holds all game screens
+int main()
+{
+    // Vector of screens, holds all game screens
     std::vector<Screens*> screens;
-    //manages which screen to run; -1 indicates close status
+    
+	// Manages which screen to run; -1 indicates close status
     int screenSelect=0;
 
-    //create game window
-    sf::RenderWindow App(sf::VideoMode(WIDTH, HEIGHT), "SPACE GAME");
+    // Create game window
+    sf::RenderWindow App(sf::VideoMode((unsigned)WIDTH, (unsigned)HEIGHT), "SPACE GAME");
     App.setFramerateLimit(FRAMERATE);
 
-    //hide mouse cursor
+    // Hide mouse cursor
     App.setMouseCursorVisible(false);
 
-    //Make screens
+    // Make main menu
     MenuScreen ms;
     screens.push_back(&ms);
+    
+	// Make game world
     World gw;
     screens.push_back(&gw);
 
-    //main game loop
-    while(screenSelect>=0){
+    // Main game loop
+    while(screenSelect>=0)
+	{
         screenSelect = screens[screenSelect]->Run(App);
     }
-    //return value set, should not happen
+    
+	// Return value set but should not happen
     return 0;
 }
